fix: `min_access_level` parameter value
Fixes #214

+++ src/gitlab/project.rs
@@ -22,12 +22,14 @@ pub struct Project {
 }
 
 impl GitLabResourceLister<Self> for Project {
+    #[expect(clippy::as_conversions, reason = "AccessLevel values are < 255")]
     fn first_url() -> String {
         format!(
             "https://{}/api/v4/projects?per_page=100&archived=false{}",
             CONFIG.connection.hostname,
             if CONFIG.owned_entities_only {
-                format!("&min_access_level={}", token::AccessLevel::Owner)
+                // `min_access_level` requires an integer, so we use `as u8` to get the numeric value
+                format!("&min_access_level={}", token::AccessLevel::Owner as u8)
             } else {
                 String::new()
             }
